processing/requirements.txt
connexion[flask,uvicorn,swagger-ui]
apscheduler
requests
PyYAML
setuptools
httpx
pykafka